tentando corrigiir pequeno bug

+++ src/Gauss/Gauss.h
@@ -9,7 +9,7 @@ class Gauss {
         Gauss(Matrix m, vector <double> b);
         
         void eliminar_gauss();
-        virtual void iterar(int i, float x) = 0;
+        virtual void iterar(int i, double x) = 0;
     protected:
         Matrix mat;
         std::vector <double> b;
